Split up worker code
