+++ include/worker/CGroup.h
@@ -1,0 +1,26 @@
+#pragma once
+
+#include <string>
+
+namespace worker {
+    const std::string BASE_CGROUP_NAME = "faasm";
+
+    enum CgroupMode {
+        cg_off, cg_on
+    };
+
+    class CGroup {
+    public:
+        explicit CGroup(const std::string &name);
+
+        void addCurrentThread();
+
+        const std::string getName();
+
+        const CgroupMode getMode();
+
+    private:
+        std::string name;
+        CgroupMode mode;
+    };
+}
