+++ include/worker/StateThread.h
@@ -1,0 +1,10 @@
+#pragma once
+
+namespace worker {
+    class StateThread {
+    public:
+        StateThread();
+
+        void run();
+    };
+}
