+++ include/worker/WorkerThreadPool.h
@@ -1,66 +1,12 @@
 #pragma once
 
-#include <scheduler/Scheduler.h>
+#include "NetworkNamespace.h"
+
 #include <wasm/wasm.h>
-
-#include <string>
-#include <exception>
-#include <tuple>
-#include <thread>
-
-#include <proto/faasm.pb.h>
-#include <queue>
-
+#include <util/func.h>
+#include <scheduler/Scheduler.h>
 
 namespace worker {
-    const std::string BASE_NETNS_NAME = "faasmns";
-    const std::string BASE_CGROUP_NAME = "faasm";
-
-    /** CGroup management */
-    enum CgroupMode {
-        cg_off, cg_on
-    };
-
-    class CGroup {
-    public:
-        explicit CGroup(const std::string &name);
-
-        void addCurrentThread();
-
-        const std::string getName();
-
-        const CgroupMode getMode();
-
-    private:
-        std::string name;
-        CgroupMode mode;
-    };
-
-    /** Network isolation */
-    enum NetworkIsolationMode {
-        ns_off, ns_on
-    };
-
-    class NetworkNamespace {
-    public:
-        explicit NetworkNamespace(const std::string &name);
-
-        void addCurrentThread();
-
-        void removeCurrentThread();
-
-        const std::string getName();
-
-        const NetworkIsolationMode getMode();
-
-    private:
-        std::string name;
-        NetworkIsolationMode mode;
-    };
-
-    /**
-     * Worker threads
-     */
     class WorkerThreadPool;
 
     class WorkerThread {
@@ -138,13 +84,5 @@ namespace worker {
         util::TokenPool prewarmTokenPool;
 
         scheduler::Scheduler &scheduler;
-    };
-
-
-    class StateThread {
-    public:
-        StateThread();
-
-        void run();
     };
 }
