+++ src/worker/CGroup.cpp
@@ -1,6 +1,9 @@
-#include "worker/worker.h"
+#include "CGroup.h"
 
 #include <util/environment.h>
+#include <util/logging.h>
+
+#include <mutex>
 
 #include <syscall.h>
 #include <sys/types.h>
