+++ src/worker/CMakeLists.txt
@@ -3,14 +3,17 @@ include_directories(
         ${FAASM_WAVM_SOURCE_DIR}
 )
 
+file(GLOB HEADERS "${FAASM_INCLUDE_DIR}/worker/*.h")
+
 add_library(worker_lib
-        cgroup.cpp
-        network.cpp
-        pool.cpp
-        ${FAASM_INCLUDE_DIR}/worker/worker.h
+        CGroup.cpp
+        NetworkNamespace.cpp
+        StateThread.cpp
+        WorkerThreadPool.cpp
+        ${HEADERS}
         )
 
 target_link_libraries(worker_lib wasm scheduler)
 
-add_executable(worker worker.cpp)
+add_executable(worker worker_main.cpp)
 target_link_libraries(worker worker_lib)
