+++ test/worker/test_cgroup.cpp
@@ -3,7 +3,7 @@
 #include <util/environment.h>
 #include <util/files.h>
 
-#include <worker/worker.h>
+#include <worker/CGroup.h>
 
 #include <boost/filesystem.hpp>
 #include <boost/algorithm/string.hpp>
